+++ structure/user.go
@@ -17,10 +17,10 @@ type UsersRequest struct {
 }
 
 type AdminUserShort struct {
-	TableName string `sql:"admin_service.users" json:"-"`
 	Image     string `json:"image"`
 	FirstName string `json:"firstName"`
 	LastName  string `json:"lastName"`
 	Email     string `json:"email" valid:"required~Required"`
 	Phone     string `json:"phone"`
+	Role      string `json:"role"`
 }